requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.3
idna==3.2
mysql-connector==2.2.9
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tweepy==3.10.0
urllib3==1.26.6
